+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Add additional checks to `/api/system/healthcheck`
 - Fix disabling asset picker in wysiwyg field
 - Improve fields projection (eg nested fields) when using MongoLite as data storage
+- Use AG Grid instead of Tabulator.js as a foundation for the table field
 
 ## 2.5.2 (2023-05-18)